move elimination machinery to GaussianHybridFactorGraph
