+++ gtsam/hybrid/HybridFactorGraph.cpp
@@ -129,58 +129,4 @@ DecisionTreeFactor::shared_ptr HybridFactorGraph::toDecisionTreeFactor() const {
   return boost::make_shared<DecisionTreeFactor>(discreteKeys(), gfgdt);
 }
 
-// The function type that does a single elimination step on a variable.
-pair<GaussianMixture::shared_ptr, boost::shared_ptr<Factor>> EliminateHybrid(
-    const GaussianHybridFactorGraph& factors, const Ordering& ordering) {
-  // STEP 1: SUM
-  // Create a new decision tree with all factors gathered at leaves.
-  Sum sum = factors.sum();
-
-  // STEP 1: ELIMINATE
-  // Eliminate each sum using conventional Cholesky:
-  // We can use this by creating a *new* decision tree:
-  using Pair = GaussianFactorGraph::EliminationResult;
-
-  KeyVector keys;
-  KeyVector separatorKeys;  // Do with optional?
-  auto eliminate = [&](const GaussianFactorGraph& graph) {
-    auto result = EliminatePreferCholesky(graph, ordering);
-    if (keys.size() == 0) keys = result.first->keys();
-    if (separatorKeys.size() == 0) separatorKeys = result.second->keys();
-    return result;
-  };
-  DecisionTree<Key, Pair> eliminationResults(sum, eliminate);
-
-  // STEP 3: Create result
-  auto pair = unzip(eliminationResults);
-  const GaussianMixture::Conditionals& conditionals = pair.first;
-  const DCGaussianMixtureFactor::Factors& separatorFactors = pair.second;
-
-  const DiscreteKeys discreteKeys = factors.discreteKeys();
-
-  // Create the GaussianMixture from the conditionals
-  auto conditional =
-      boost::make_shared<GaussianMixture>(keys, discreteKeys, conditionals);
-
-  // If there are no more continuous parents, then we should create here a
-  // DiscreteFactor, with the error for each discrete choice.
-  if (separatorKeys.size() == 0) {
-    VectorValues empty_values;
-    auto factorError = [&](const GaussianFactor::shared_ptr& factor) {
-      return exp(-factor->error(empty_values));
-    };
-    DecisionTree<Key, double> fdt(separatorFactors, factorError);
-    auto discreteFactor =
-        boost::make_shared<DecisionTreeFactor>(factors.discreteKeys(), fdt);
-
-    return {conditional, discreteFactor};
-
-  } else {
-    // Create a resulting DCGaussianMixture on the separator.
-    auto factor = boost::make_shared<DCGaussianMixtureFactor>(
-        separatorKeys, discreteKeys, separatorFactors);
-    return {conditional, factor};
-  }
-}
-
 }  // namespace gtsam
